# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  commute.CommuteResponse:
    properties:
      commute_time:
        example: 30
        type: integer
      destination_branch:
        example: 2
        type: integer
      source_branch:
        example: 1
        type: integer
    type: object
  commute.UpdateCommuteRequest:
    properties:
      commute_time:
        example: 30
        type: integer
    required:
    - commute_time
    type: object
...
  scheduling.ErrorResponse:
    properties:
      error:
        example: failed to retrieve teachers
        type: string
    type: object
...
info:
  contact: {}
  description: API for teacher availability management and bookings
  title: Converge API
  version: "1.0"
paths:
...
  /commute:
    get:
      description: Returns the commute time in minutes. Without query params, returns
        the default commute time. With both source_branch and destination_branch,
        returns 0 when they are equal, otherwise the default.
      parameters:
      - description: Source branch ID
        in: query
        name: source_branch
        type: integer
      - description: Destination branch ID
        in: query
        name: destination_branch
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commute.CommuteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get commute time between two branches
      tags:
      - commute
    patch:
      consumes:
      - application/json
      description: Sets the commute time in minutes applied between different branches.
        Must be non-negative.
      parameters:
      - description: New commute time
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/commute.UpdateCommuteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/commute.CommuteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/scheduling.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update the global commute time
      tags:
      - commute
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT from /login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"